refactor: move remote start api endpoint to new structure.
